build: use shared build scripts
